refactor(app): finish Flame wave-10 grandfather file splits (#4224)
Split four grandfathered Flame modules below the 400 physical-line cap and
clear the shrink-only allowlist. Structural-only; no render semantics change.
